Fix scope checking for HerderStatusMonitor
The HerderStatusMonitor only checks whether the current worker ID is the
active worker when monitoring task status. At the connector level,
there's no checking against current worker ID, so each instance in the
cluster will report metrics and potentially attempt a restart for all
connectors.

Before performing Connecter processing, check that the current worker is
the assigned worker for the connector to ensure that metrics and
restarts are only reported once.