Handle tombstone records in offset tool
In order to support the ability to delete offset entries in the Mirus configured offset topic, the offset tool should be able to handle offset entries of null. Kafka will eventually delete records in a compact topic that has the same key value of a tombstone record. This deletion doesn't happen instantly and depends on different configuration settings.
An outcome of this feature of handling null offsets is the ability of a consumer application, such as an offset checker, to have offsets of partitions of a deleted topic marked with a tombstone record and therefore filtered out.