Continue running KafkaMonitor after an exception is encountered
The KafkaMonitor thread needs to always continue running and monitoring
for partition changes in the source and destination clusters in order
for it to trigger task reconfiguration. Catch exceptions inside the
while(true) loop so the thread continues after retriable exceptions.

Also ensure that the shutDownLatch is checked during exponential backoff
retries, so the thread shuts down cleanly even in an error state.