Skip records that have already been consumed by a task (#61)
This works around the KIP-320 bug present up to Kafka 2.3.0 where the current consumer offsets
can become invalid after partition leadership changes. In this case the same records 
causing many duplicates to be produced.

Note that this change does not need to be applied to master as it uses Kafka 2.3.1.